od5.0: Fixed double to float conversion warning
Summary:
src\AttributeEngine\attribdatapack.cc(997) : warning C4244: 'argument' : conversion from 'double' to 'const float', possible loss of data

src\AttributeEngine\attribdatapack.cc(1010) : warning C4244: 'argument' : conversion from 'double' to 'const float', possible loss of data

Test Plan: Code compiles

Reviewers: henrique

Reviewed By: henrique

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D7308

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38493 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/AttributeEngine/attribdatapack.cc
@@ -683,7 +683,7 @@ FlatRdmTrcsDataPack::FlatRdmTrcsDataPack( DescID did, const SeisTrcBuf& sb,
     refnrs1_.setParam( this, new TypeSet<float> );
     if ( path )
 	path_ = new TypeSet<BinID>(*path);
-    
+
     SamplingData<float>& sd = const_cast<SamplingData<float>& >(samplingdata_);
     const SeisTrc* firsttrc = sb.get( 0 );
     if ( firsttrc )
@@ -993,7 +993,7 @@ bool FlatDataPackZAxisTransformer::doFinish( bool success )
 	{
 	    TypeSet<float> refnrs;
 	    for ( int idx=0; idx<dataarr->trcinfoset_.size(); idx++ )
-		refnrs += dataarr->trcinfoset_[idx]->getValue(
+		refnrs += (float)dataarr->trcinfoset_[idx]->getValue(
 						SeisTrcInfo::RefNr );
 	    outputf2ddp->setRefNrs( refnrs );
 	}
@@ -1006,7 +1006,7 @@ bool FlatDataPackZAxisTransformer::doFinish( bool success )
 			const_cast<TypeSet<BinID>* >(dprdm->pathBIDs()) );
 	TypeSet<float> refnrs;
 	for ( int idx=0; idx<dprdm->seisBuf().size(); idx++ )
-	    refnrs += dprdm->seisBuf().get(idx)->info().getValue(
+	    refnrs += (float)dprdm->seisBuf().get(idx)->info().getValue(
 						SeisTrcInfo::RefNr );
 	outputdprdm->setRefNrs( refnrs );
 	outputdp = outputdprdm;
